requirements.txt
fastapi
uvicorn
scikit-learn
pillow
numpy